+++ src/NowoFormKitBundle.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Nowo\FormKitBundle;
 
+use Nowo\FormKitBundle\DependencyInjection\Compiler\TwigPathsPass;
 use Nowo\FormKitBundle\DependencyInjection\FormKitExtension;
 use Nowo\FormKitBundle\DependencyInjection\FormOptionsMergerInjectorCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -30,6 +31,8 @@ class NowoFormKitBundle extends Bundle
     public function build(ContainerBuilder $container): void
     {
         parent::build($container);
+
+        $container->addCompilerPass(new TwigPathsPass());
         $container->addCompilerPass(new FormOptionsMergerInjectorCompilerPass());
     }
 }